requirements.txt
openai
requests
Pillow
tqdm
transformers
torch
numpy
matplotlib
scikit-learn
opencv-python
